integrate examples into a single file
